requirements.txt
langchain==0.3.13
streamlit==1.41.1
psycopg2==2.9.10
langchain-ollama==0.2.2
pillow==11.0.0
langchain-community==0.3.13
